backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.10
pyyaml==6.0.2
pymupdf==1.24.10
